refactor: 폴더 경로가 themeParks로 되어있던 것 themeparks로 수정
