+++ src/main/java/com/beyond/easycheck/attractions/application/service/AttractionService.java
@@ -4,9 +4,9 @@ import com.beyond.easycheck.attractions.exception.AttractionMessageType;
 import com.beyond.easycheck.attractions.infrastructure.entity.AttractionEntity;
 import com.beyond.easycheck.attractions.infrastructure.repository.AttractionRepository;
 import com.beyond.easycheck.common.exception.EasyCheckException;
-import com.beyond.easycheck.themeParks.exception.ThemeParkMessageType;
-import com.beyond.easycheck.themeParks.infrastructure.persistence.entity.ThemeParkEntity;
-import com.beyond.easycheck.themeParks.infrastructure.persistence.repository.ThemeParkRepository;
+import com.beyond.easycheck.themeparks.exception.ThemeParkMessageType;
+import com.beyond.easycheck.themeparks.infrastructure.entity.ThemeParkEntity;
+import com.beyond.easycheck.themeparks.infrastructure.repository.ThemeParkRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
