+++ src/main/java/com/beyond/easycheck/attractions/infrastructure/entity/AttractionEntity.java
@@ -1,7 +1,7 @@
 package com.beyond.easycheck.attractions.infrastructure.entity;
 
 import com.beyond.easycheck.attractions.application.service.AttractionOperationUseCase.AttractionCreateCommand;
-import com.beyond.easycheck.themeParks.infrastructure.persistence.entity.ThemeParkEntity;
+import com.beyond.easycheck.themeparks.infrastructure.entity.ThemeParkEntity;
 import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
