+++ src/main/java/com/beyond/easycheck/themeParks/infrastructure/entity/ThemeParkEntity.java
@@ -1,8 +1,8 @@
-package com.beyond.easycheck.themeParks.infrastructure.persistence.entity;
+package com.beyond.easycheck.themeparks.infrastructure.entity;
 
 
 import com.beyond.easycheck.common.entity.BaseTimeEntity;
-import com.beyond.easycheck.themeParks.application.service.ThemeParkOperationUseCase.ThemeParkCreateCommand;
+import com.beyond.easycheck.themeparks.application.service.ThemeParkOperationUseCase.ThemeParkCreateCommand;
 import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
