+++ src/main/java/com/beyond/easycheck/themeParks/infrastructure/repository/ThemeParkRepository.java
@@ -1,6 +1,6 @@
-package com.beyond.easycheck.themeParks.infrastructure.persistence.repository;
+package com.beyond.easycheck.themeparks.infrastructure.repository;
 
-import com.beyond.easycheck.themeParks.infrastructure.persistence.entity.ThemeParkEntity;
+import com.beyond.easycheck.themeparks.infrastructure.entity.ThemeParkEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface ThemeParkRepository extends JpaRepository<ThemeParkEntity, Long> {
